Restructuring and filling in. Fetid! How dare they present this to me?!? Foul! They hide their finest bean... Prepare the attack!!!

+++ java/net/RPower/RPowermod/machines/power/cable/BlockFluxCableBasic.java
@@ -14,7 +14,7 @@ public class BlockFluxCableBasic extends Block{
 	
 	@Override
 	public TileEntity createTileEntity(World world, int metadata) {
-		return (TileEntity) new TileEntityFluxCable(32);
+		return (TileEntity) new TileEntityFluxCable((byte)32);
 	}
 
 }
